完善帐号维护功能

+++ src/main/java/com/hb/weibo/repository/WeiboAccountDao.java
@@ -14,5 +14,5 @@ import com.hb.weibo.entity.WeiboAccount;
 public interface WeiboAccountDao extends PagingAndSortingRepository<WeiboAccount, Long> {
 	@Modifying
 	@Query("update WeiboAccount weiboAccount set weiboAccount.nick=?1 where weiboAccount.type=?2 and weiboAccount.username=?3")
-	void updateNick(String nick, Long type, String username);
+	void updateNick(String nick, int type, String username);
 }
